Added most of the functionalities
The add family feature is done, along with the delete family member and varioud listing. An update to the terminal gui so its more pleasing to look at and added a new function "System("CLR")" in void cleanup to clean up the terminal after almost everyuse for a clean layout.

Im looking for a better way to make the gui look more appeling in the terminal.