Added and worked on more functions
The base root program fully works. The only thing that throws it off right now is the uses of .'s in the names. So anyone that a decent of their family using the same name can't but .Jr .Sr because it will crash the program. Bug for this will be fixed at a later time. The 2nd menu for the program has been added. This leads up to what stuff can be added.